added new sax-based mzml reader (works with prism)

+++ src/main/java/edu/washington/gs/maccoss/encyclopedia/filereaders/MzmlToDIAConverter.java
@@ -24,29 +24,38 @@ public class MzmlToDIAConverter {
 	public static final String MZML_EXTENSION=".mzml";
 
 	public static void main(String[] args) throws IOException {
+		boolean copy=false;
+		
 		HashMap<String, String> paramMap=PecanParameterParser.getDefaultParameters();
 		paramMap.put("-acquisition", "DIA"); // NON-OVERLAPPING!
 		
 		SearchParameters parameters=PecanParameterParser.parseParameters(paramMap);
 		System.out.println(parameters);
 		
-		File dir=new File("/Volumes/WorkingDisk/bruker/second_try/");
+		File dir=new File("/Volumes/BriansSSD/overlapping/prism/");
 		
 		File[] files=dir.listFiles(new SimpleFilenameFilter(MZML_EXTENSION));
 		for (int i=0; i<files.length; i++) {
 			System.out.println((i+1)+" / "+files.length+"\t Copying "+files[i].getName()+"...");
 
-			File f=File.createTempFile(files[i].getName(), MZML_EXTENSION);
-			f.deleteOnExit();
-			Files.copy(files[i].toPath(), f.toPath(), StandardCopyOption.REPLACE_EXISTING);
+			File f;
+			if (copy) {
+				f=File.createTempFile(files[i].getName(), MZML_EXTENSION);
+				f.deleteOnExit();
+				Files.copy(files[i].toPath(), f.toPath(), StandardCopyOption.REPLACE_EXISTING);
+			} else {
+				f=files[i];
+			}
 			
 			Long time=System.currentTimeMillis();
 			File diaFile=new File(files[i].getAbsolutePath().substring(0, files[i].getAbsolutePath().lastIndexOf('.'))+StripeFile.DIA_EXTENSION);
 			System.out.println("Converting to "+diaFile.getAbsolutePath());
 			
-			convert(f, diaFile, parameters);
-			
-			f.delete();
+			convertSAX(f, diaFile, parameters);
+
+			if (copy) {
+				f.delete();
+			}
 			System.out.println("Total time: "+(System.currentTimeMillis()-time)/1000f+" seconds");
 		}
 	}
@@ -70,7 +79,7 @@ public class MzmlToDIAConverter {
 		
 		// otherwise check for MZML and convert
 		if (f.getName().toLowerCase().endsWith(MZML_EXTENSION)) {
-			return convert(f, diaFile, parameters);
+			return convertSAX(f, diaFile, parameters);
 		} else {
 			throw new EncyclopediaException("Can't read file type "+f.getAbsolutePath());
 		}
@@ -92,7 +101,92 @@ public class MzmlToDIAConverter {
 		}
 	}
 
-	static StripeFileInterface convert(File mzMLFile, File diaFile, SearchParameters parameters) {
+	static StripeFileInterface convertSAX(File mzMLFile, File diaFile, SearchParameters parameters) {
+		try {
+			Logger.logLine("Indexing "+mzMLFile.getName()+" ...");
+			StripeFile stripeFile=new StripeFile();
+			stripeFile.openFile();
+
+			BlockingQueue<MzmlBlock> mzmlBlockQueue=new ArrayBlockingQueue<MzmlBlock>(1);
+			MzmlToDIASAXProducer producer=new MzmlToDIASAXProducer(mzMLFile, mzmlBlockQueue, parameters);
+			
+			// will be populated after we join back up. Since we're not looking
+			// at it until after the join, we're safe to not have to worry about
+			// concurrency.
+			HashMap<Range, TFloatArrayList> retentionTimesByStripe=producer.getRetentionTimesByStripe();
+			Thread[] threads;
+			
+			if (parameters.isDeconvoluteOverlappingWindows()) {
+				BlockingQueue<MzmlBlock> deconvolutionBlockQueue=new ArrayBlockingQueue<MzmlBlock>(1);
+				OverlapDeconvoluter deconvoluter=new OverlapDeconvoluter(parameters.getFragmentTolerance(), mzmlBlockQueue, deconvolutionBlockQueue);
+				retentionTimesByStripe=deconvoluter.getRetentionTimesByStripe();
+				MzmlToDIAConsumer consumer=new MzmlToDIAConsumer(deconvolutionBlockQueue, stripeFile);
+
+				Logger.logLine("Converting "+mzMLFile.getName()+" ...");
+				Thread producerThread=new Thread(producer);
+				Thread deconvoluterThread=new Thread(deconvoluter);
+				Thread consumerThread=new Thread(consumer);
+
+				threads=new Thread[] {producerThread, deconvoluterThread, consumerThread};
+				
+			} else {
+				MzmlToDIAConsumer consumer=new MzmlToDIAConsumer(mzmlBlockQueue, stripeFile);
+
+				Logger.logLine("Converting "+mzMLFile.getName()+" ...");
+				Thread producerThread=new Thread(producer);
+				Thread consumerThread=new Thread(consumer);
+
+				threads=new Thread[] {producerThread, consumerThread};
+			}
+			
+			for (int i=0; i<threads.length; i++) {
+				threads[i].start();
+			}
+
+			try {
+				for (int i=0; i<threads.length; i++) {
+					threads[i].join();
+				}
+
+				stripeFile.setFileName(mzMLFile.getName(), producer.getMzMLID(), mzMLFile.getAbsolutePath());
+
+				Logger.logLine("Finalizing "+diaFile.getName()+" ...");
+				HashMap<Range, Float> dutyCycleMap=new HashMap<Range, Float>();
+				for (Entry<Range, TFloatArrayList> entry : retentionTimesByStripe.entrySet()) {
+					Range range=entry.getKey();
+					TFloatArrayList rts=entry.getValue();
+					float[] deltas=General.firstDerivative(rts.toArray());
+					float averageDutyCycle=General.mean(deltas);
+					dutyCycleMap.put(range, averageDutyCycle);
+				}
+				if (!parameters.isDDA()) {
+					stripeFile.setRanges(dutyCycleMap);
+				}
+
+				stripeFile.saveAsFile(diaFile);
+				stripeFile.close();
+				
+				stripeFile=new StripeFile();
+				stripeFile.openFile(diaFile);
+				Logger.logLine("Finished writing "+diaFile.getName()+"!");
+
+				return stripeFile;
+
+			} catch (InterruptedException ie) {
+				Logger.errorLine("DIA writing interrupted!");
+				Logger.errorException(ie);
+				return null;
+			}
+
+		} catch (IOException ioe) {
+			throw new EncyclopediaException("DIA writing IO error!", ioe);
+		} catch (SQLException sqle) {
+			sqle.printStackTrace();
+			throw new EncyclopediaException("DIA writing SQL error!", sqle);
+		}
+	}
+
+	static StripeFileInterface convertJMZML(File mzMLFile, File diaFile, SearchParameters parameters) {
 		try {
 			Logger.logLine("Indexing "+mzMLFile.getName()+" ...");
 			StripeFile stripeFile=new StripeFile();
